Fixes from the second rebase; new modal
